# docs/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Hefesto API
  version: "2.0.0"
  description: |
    API REST da plataforma Hefesto — agentes de IA para prospecção e entrega de sites.
    Autenticação: Bearer JWT RS256.
    Todos os endpoints retornam envelope { data, meta } ou { errors }.
  contact:
    name: Hefesto
servers:
  - url: http://localhost:3001
    description: Desenvolvimento local
  - url: https://api.seudominio.com
    description: Produção

# ===========================================================================
# SECURITY
# ===========================================================================
...
security:
  - BearerAuth: []

# ===========================================================================
# COMPONENTS — Schemas reutilizáveis
# ===========================================================================
components:
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string, example: "Email inválido" }
        field: { type: string, example: "email" }
        requestId: { type: string, example: "01HZ..." }
    ApiErrorResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items: { $ref: "#/components/schemas/ApiError" }

    # --- Paginação ---
    PaginationMeta:
      type: object
      properties:
        cursor:
          type: object
          properties:
            next: { type: string, nullable: true }
            prev: { type: string, nullable: true }
        total: { type: integer }
        limit: { type: integer }

    # --- Auth ---
...
    LeadStatus:
      type: string
      enum: [NEW, CONTACTED, QUALIFIED, NEGOTIATING, CONVERTED, LOST]
    LeadSource:
      type: string
      enum: [MANUAL, GOOGLE_MAPS, SCRAPED, REFERRAL, APOLLO]
    MessageChannel:
      type: string
      enum: [WHATSAPP, TELEGRAM, EMAIL, INTERNAL]
    EnrichmentData:
      type: object
      properties:
        cnpj: { type: string, nullable: true }
        cnpjStatus:
          {
            type: string,
            enum: [ATIVA, SUSPENSA, INAPTA, BAIXADA, NULA],
            nullable: true,
          }
        yearsInBusiness: { type: integer, nullable: true }
        googleMapsPlaceId: { type: string, nullable: true }
        googleRating: { type: number, nullable: true }
        googleReviewsCount: { type: integer, nullable: true }
        hasWebsite: { type: boolean }
        websiteQualityHint:
          {
            type: string,
            enum: [modern, outdated, mobile_broken, none],
            nullable: true,
          }
        neighborhood: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
    Lead:
      type: object
      properties:
        id: { type: string, format: uuid }
        contactName: { type: string }
        businessName: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, format: email, nullable: true }
        source: { $ref: "#/components/schemas/LeadSource" }
        status: { $ref: "#/components/schemas/LeadStatus" }
        preferredChannel: { $ref: "#/components/schemas/MessageChannel" }
        qualificationScore:
          { type: integer, minimum: 0, maximum: 100, nullable: true }
        enrichmentData: { $ref: "#/components/schemas/EnrichmentData" }
        followUpCount: { type: integer }
        nextFollowUpAt: { type: string, format: date-time, nullable: true }
        assignedAgentId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateLeadRequest:
      type: object
      required: [contactName, source]
      properties:
        contactName: { type: string, minLength: 2, maxLength: 100 }
        businessName: { type: string, maxLength: 200 }
        contactPhone: { type: string, pattern: '^\+55\d{10,11}$' }
        contactEmail: { type: string, format: email }
        source: { $ref: "#/components/schemas/LeadSource" }
        preferredChannel: { $ref: "#/components/schemas/MessageChannel" }
        # Ao menos um de contactPhone ou contactEmail é obrigatório

    # --- HITL ---
...
  responses:
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorResponse" }
    ValidationError:
      description: Dados inválidos
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorResponse" }
...
paths:
...
  /api/v1/leads:
    get:
      tags: [Leads]
      summary: Listar leads com filtros
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/LeadStatus" }
        - name: source
          in: query
          schema: { $ref: "#/components/schemas/LeadSource" }
        - name: channel
          in: query
          schema: { $ref: "#/components/schemas/MessageChannel" }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista de leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Lead" }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }

    post:
      tags: [Leads]
      summary: Criar lead manualmente
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLeadRequest" }
      responses:
        "201":
          description: Lead criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Lead" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /api/v1/leads/{id}:
    get:
      tags: [Leads]
      summary: Buscar lead por ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Lead com todos os dados
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Lead" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --- HITL ---
...